Implement bot.accounts
Uses IRCv3.1 `account-notify` and `extended-join` to ensure that it's up to date.
Should be mostly foolproof, and should stay up to date as much as possible. When a user parts from a channel, or quits IRC, they will be removed from account lists. Unfortunately, accounts will only be tracked for users that share a channel with Sopel, if a user is not in one of Sopel's channels, they will not be tracked.